requirements.txt
bidict==0.21.4
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-SocketIO==5.3.6
h11==0.13.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
pep517==0.13.0
pip-tools==6.4.0
pyalsaaudio==0.10.0
PySimpleGUI==4.60.5
PySimpleGUIWeb==0.39.0
python-engineio==4.7.1
python-socketio==5.9.0
remi==2020.3.10
simple-websocket==1.0.0
tomli==1.2.3
typing_extensions==4.1.1
Wave==0.0.2
Werkzeug==2.0.3
wsproto==1.0.0
zipp==3.6.0
